requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
lxml[html_clean]
newspaper3k==0.2.8
PyPDF2==3.0.1
python-dotenv==1.0.0
tqdm==4.66.1
selenium==4.16.0
webdriver-manager==4.0.1
playwright==1.41.2
cloudscraper==1.2.71
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
feedparser==6.0.10
fastapi
pydantic
uvicorn
